fix: 리포트 페이지 다국어 작업을 위한 데이터 수정
