+++ src/utils/calculate-period.ts
@@ -1,7 +1,7 @@
-import { TMappedPeriod } from "@/types/StockDataType";
+import { TPeriod } from "@/stores/useReportStore";
 
 // create startDateTime and endDateTime period calculation ex) 202308010000, 202407181725
-const calculatePeriod = (period: keyof TMappedPeriod) => {
+const calculatePeriod = (period: TPeriod) => {
   const formatDateTime = (date: Date) => {
     const year = date.getFullYear();
     const month = String(date.getMonth() + 1).padStart(2, "0");
@@ -16,19 +16,19 @@ const calculatePeriod = (period: keyof TMappedPeriod) => {
   let endDateTime = formatDateTime(currentDate);
 
   switch (period) {
-    case "일":
+    case "day":
       startDateTime = formatDateTime(new Date(new Date().setMonth(currentDate.getMonth() - 3)));
       break;
-    case "주":
+    case "week":
       startDateTime = formatDateTime(new Date(new Date().setFullYear(currentDate.getFullYear() - 1)));
       break;
-    case "월":
+    case "month":
       startDateTime = formatDateTime(new Date(new Date().setFullYear(currentDate.getFullYear() - 3)));
       break;
-    case "분기":
+    case "quarter":
       startDateTime = formatDateTime(new Date(new Date().setFullYear(currentDate.getFullYear() - 5)));
       break;
-    case "년":
+    case "year":
       startDateTime = formatDateTime(new Date(new Date().setFullYear(currentDate.getFullYear() - 7)));
       break;
     default:
